@@ -2,8 +2,8 @@ package rw.solution.easy.dental.model;
 
 import java.io.Serializable;
 
-import javax.persistence.Column;
-import javax.persistence.Embeddable;
+import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
 
 @Embeddable
 public class TratamentoProcedimentoPK implements Serializable{
