@@ -13,8 +13,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import rw.solution.easy.dental.model.dto.AgendaDto;
 import rw.solution.easy.dental.service.AgendaService;
 import rw.solution.easy.dental.util.LogUtil;
@@ -22,7 +22,7 @@ import rw.solution.easy.dental.util.LogUtil;
 @CrossOrigin
 @RestController
 @RequestMapping("/agenda/v1/")
-@Api(tags = {"Agenda"})
+@Tag(name = "Agenda", description = "Métodos da agenda")
 public class AgendaController implements Serializable {
 
 	/**
@@ -35,7 +35,7 @@ public class AgendaController implements Serializable {
 	@Autowired
 	private AgendaService service;
 	
-	@ApiOperation(value = "Recupera os agendamentos", response = AgendaDto.class)
+	@Operation(summary = "Recupera os agendamentos")
 	@GetMapping(value = "/{customer}", produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity<AgendaDto> getAgendamentos(@PathVariable(required=true) Long customer) {
 		
